requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-dotenv
azure-cosmos
azure-monitor-opentelemetry
